requirements.txt
tqdm
logfire
loguru
python-dotenv
